// amplify/graphql/API.ts
export enum ModelSortDirection {
  ASC = "ASC",
  DESC = "DESC",
}
// ...
export type GetAircraftQueryVariables = {
  id: string,
};

export type GetAircraftQuery = {
  getAircraft?:  {
    __typename: "Aircraft",
    archived: boolean,
    createdAt: string,
    description?: string | null,
    document?:  {
      __typename: "ModelAircraftDocumentConnection",
      nextToken?: string | null,
    } | null,
    id: string,
    name: string,
    notices?:  {
      __typename: "ModelAircraftNoticeConnection",
      nextToken?: string | null,
    } | null,
    staff?:  {
      __typename: "ModelAircraftStaffConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
  } | null,
};

export type GetAircraftDocumentQueryVariables = {
  id: string,
};

export type GetAircraftDocumentQuery = {
  getAircraftDocument?:  {
    __typename: "AircraftDocument",
    aircraft?:  {
      __typename: "Aircraft",
      archived: boolean,
      createdAt: string,
      description?: string | null,
      id: string,
      name: string,
      updatedAt: string,
    } | null,
    aircraftId: string,
    createdAt: string,
    document?:  {
      __typename: "Document",
      archived: boolean,
      createdAt: string,
      id: string,
      name: string,
      staffId?: string | null,
      subcategoryId?: string | null,
      updatedAt: string,
    } | null,
    documentId: string,
    id: string,
    updatedAt: string,
  } | null,
};

export type GetAircraftNoticeQueryVariables = {
  id: string,
};

export type GetAircraftNoticeQuery = {
  getAircraftNotice?:  {
    __typename: "AircraftNotice",
    aircraft?:  {
      __typename: "Aircraft",
      archived: boolean,
      createdAt: string,
      description?: string | null,
      id: string,
      name: string,
      updatedAt: string,
    } | null,
    aircraftId: string,
    createdAt: string,
    id: string,
    notice?:  {
      __typename: "Notice",
      archived: boolean,
      createdAt: string,
      deadline_at?: string | null,
      details: string,
      id: string,
      noticed_at?: string | null,
      staffId?: string | null,
      status?: NoticeStatus | null,
      subject: string,
      type?: NoticeType | null,
      updatedAt: string,
    } | null,
    noticeId: string,
    updatedAt: string,
  } | null,
};

export type GetAircraftStaffQueryVariables = {
  id: string,
};

export type GetAircraftStaffQuery = {
  getAircraftStaff?:  {
    __typename: "AircraftStaff",
    aircraft?:  {
      __typename: "Aircraft",
      archived: boolean,
      createdAt: string,
      description?: string | null,
      id: string,
      name: string,
      updatedAt: string,
    } | null,
    aircraftId: string,
    createdAt: string,
    id: string,
    staff?:  {
      __typename: "Staff",
      archived: boolean,
      createdAt: string,
      email: string,
      id: string,
      name: string,
      updatedAt: string,
    } | null,
    staffId: string,
    updatedAt: string,
  } | null,
};

export type GetCategoryQueryVariables = {
  id: string,
};

export type GetCategoryQuery = {
  getCategory?:  {
    __typename: "Category",
    archived: boolean,
    createdAt: string,
    description?: string | null,
    id: string,
    name: string,
    subcategories?:  {
      __typename: "ModelSubcategoryConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
  } | null,
};

export type GetDocumentQueryVariables = {
  id: string,
};

export type GetDocumentQuery = {
  getDocument?:  {
    __typename: "Document",
    aircraft?:  {
      __typename: "ModelAircraftDocumentConnection",
      nextToken?: string | null,
    } | null,
    archived: boolean,
    createdAt: string,
    id: string,
    name: string,
    staff?:  {
      __typename: "Staff",
      archived: boolean,
      createdAt: string,
      email: string,
      id: string,
      name: string,
      updatedAt: string,
    } | null,
    staffId?: string | null,
    subcategory?:  {
      __typename: "Subcategory",
      archived: boolean,
      categoryId: string,
      createdAt: string,
      description?: string | null,
      id: string,
      name: string,
      updatedAt: string,
    } | null,
    subcategoryId?: string | null,
    updatedAt: string,
  } | null,
};

export type GetNoticeQueryVariables = {
  id: string,
};

export type GetNoticeQuery = {
  getNotice?:  {
    __typename: "Notice",
    aircraft?:  {
      __typename: "ModelAircraftNoticeConnection",
      nextToken?: string | null,
    } | null,
    archived: boolean,
    author?:  {
      __typename: "Staff",
      archived: boolean,
      createdAt: string,
      email: string,
      id: string,
      name: string,
      updatedAt: string,
    } | null,
    createdAt: string,
    deadline_at?: string | null,
    details: string,
    documents?:  {
      __typename: "ModelNoticeDocumentConnection",
      nextToken?: string | null,
    } | null,
    id: string,
    noticed_at?: string | null,
    recipients?:  {
      __typename: "ModelNotificationConnection",
      nextToken?: string | null,
    } | null,
    staffId?: string | null,
    status?: NoticeStatus | null,
    subject: string,
    type?: NoticeType | null,
    updatedAt: string,
  } | null,
};

export type GetNoticeDocumentQueryVariables = {
  id: string,
};

export type GetNoticeDocumentQuery = {
  getNoticeDocument?:  {
    __typename: "NoticeDocument",
    createdAt: string,
    id: string,
    name: string,
    noticeId: string,
    notices?:  {
      __typename: "Notice",
      archived: boolean,
      createdAt: string,
      deadline_at?: string | null,
      details: string,
      id: string,
      noticed_at?: string | null,
      staffId?: string | null,
      status?: NoticeStatus | null,
      subject: string,
      type?: NoticeType | null,
      updatedAt: string,
    } | null,
    updatedAt: string,
  } | null,
};

export type GetNotificationQueryVariables = {
  id: string,
};

export type GetNotificationQuery = {
  getNotification?:  {
    __typename: "Notification",
    createdAt: string,
    id: string,
    notice?:  {
      __typename: "Notice",
      archived: boolean,
      createdAt: string,
      deadline_at?: string | null,
      details: string,
      id: string,
      noticed_at?: string | null,
      staffId?: string | null,
      status?: NoticeStatus | null,
      subject: string,
      type?: NoticeType | null,
      updatedAt: string,
    } | null,
    noticeId: string,
    read_at?: string | null,
    staff?:  {
      __typename: "Staff",
      archived: boolean,
      createdAt: string,
      email: string,
      id: string,
      name: string,
      updatedAt: string,
    } | null,
    staffId: string,
    updatedAt: string,
  } | null,
};

export type GetRoleQueryVariables = {
  id: string,
};

export type GetRoleQuery = {
  getRole?:  {
    __typename: "Role",
    archived: boolean,
    createdAt: string,
    description?: string | null,
    id: string,
    name: string,
    staff?:  {
      __typename: "ModelRoleStaffConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
  } | null,
};

export type GetRoleStaffQueryVariables = {
  id: string,
};

export type GetRoleStaffQuery = {
  getRoleStaff?:  {
    __typename: "RoleStaff",
    createdAt: string,
    id: string,
    role?:  {
      __typename: "Role",
      archived: boolean,
      createdAt: string,
      description?: string | null,
      id: string,
      name: string,
      updatedAt: string,
    } | null,
    roleId: string,
    staff?:  {
      __typename: "Staff",
      archived: boolean,
      createdAt: string,
      email: string,
      id: string,
      name: string,
      updatedAt: string,
    } | null,
    staffId: string,
    updatedAt: string,
  } | null,
};

export type GetStaffQueryVariables = {
  id: string,
};

export type GetStaffQuery = {
  getStaff?:  {
    __typename: "Staff",
    aircraft?:  {
      __typename: "ModelAircraftStaffConnection",
      nextToken?: string | null,
    } | null,
    archived: boolean,
    createdAt: string,
    documents?:  {
      __typename: "ModelDocumentConnection",
      nextToken?: string | null,
    } | null,
    email: string,
    id: string,
    name: string,
    notices?:  {
      __typename: "ModelNoticeConnection",
      nextToken?: string | null,
    } | null,
    notifications?:  {
      __typename: "ModelNotificationConnection",
      nextToken?: string | null,
    } | null,
    roles?:  {
      __typename: "ModelRoleStaffConnection",
      nextToken?: string | null,
    } | null,
    subcategories?:  {
      __typename: "ModelStaffSubcategoryConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
  } | null,
};

export type GetStaffSubcategoryQueryVariables = {
  id: string,
};

export type GetStaffSubcategoryQuery = {
  getStaffSubcategory?:  {
    __typename: "StaffSubcategory",
    accessLevel: number,
    createdAt: string,
    id: string,
    staff?:  {
      __typename: "Staff",
      archived: boolean,
      createdAt: string,
      email: string,
      id: string,
      name: string,
      updatedAt: string,
    } | null,
    staffId: string,
    subcategory?:  {
      __typename: "Subcategory",
      archived: boolean,
      categoryId: string,
      createdAt: string,
      description?: string | null,
      id: string,
      name: string,
      updatedAt: string,
    } | null,
    subcategoryId: string,
    updatedAt: string,
  } | null,
};

export type GetSubcategoryQueryVariables = {
  id: string,
};

export type GetSubcategoryQuery = {
  getSubcategory?:  {
    __typename: "Subcategory",
    archived: boolean,
    category?:  {
      __typename: "Category",
      archived: boolean,
      createdAt: string,
      description?: string | null,
      id: string,
      name: string,
      updatedAt: string,
    } | null,
    categoryId: string,
    createdAt: string,
    description?: string | null,
    documents?:  {
      __typename: "ModelDocumentConnection",
      nextToken?: string | null,
    } | null,
    id: string,
    name: string,
    staff?:  {
      __typename: "ModelStaffSubcategoryConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
  } | null,
};

export type ListAircraftQueryVariables = {
  filter?: ModelAircraftFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListAircraftQuery = {
  listAircraft?:  {
    __typename: "ModelAircraftConnection",
    items:  Array< {
      __typename: "Aircraft",
      archived: boolean,
      createdAt: string,
      description?: string | null,
      id: string,
      name: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListAircraftByNameQueryVariables = {
  filter?: ModelAircraftFilterInput | null,
  limit?: number | null,
  name: string,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
};

export type ListAircraftByNameQuery = {
  listAircraftByName?:  {
    __typename: "ModelAircraftConnection",
    items:  Array< {
      __typename: "Aircraft",
      archived: boolean,
      createdAt: string,
      description?: string | null,
      id: string,
      name: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListAircraftDocumentsQueryVariables = {
  filter?: ModelAircraftDocumentFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListAircraftDocumentsQuery = {
  listAircraftDocuments?:  {
    __typename: "ModelAircraftDocumentConnection",
    items:  Array< {
      __typename: "AircraftDocument",
      aircraftId: string,
      createdAt: string,
      documentId: string,
      id: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListAircraftNoticesQueryVariables = {
  filter?: ModelAircraftNoticeFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};